+++ pytest.ini
@@ -14,6 +14,7 @@ filterwarnings =
 # --last-failed to run only the tests which failed on the previous run
 # --show-capture no/stdout/stderr/both
 addopts = 
+    --keep-duplicates
     -p julia.pytestplugin 
     -vv 
     --ignore test/regression
